Cloud-Backend/requirements.txt
flask
tensorflow
scipy
scikit-learn
matplotlib
pillow
opencv-python
numpy
flask-cors
